Webmirror: changes for filename rewriting

+++ Applications/Webmirror.php
@@ -15,9 +15,10 @@ use Wa72\Spider\Core\HttpClientResponseEvent;
 class Webmirror extends AbstractSpider
 {
     protected string $output_dir;
-    protected string $link_prefix = '';
+    protected string $path_prefix = '';
     protected array $additional_urls = [];
     protected string $hostname;
+    protected array $hashes;
 
     /**
      * Functions that are called before saving response body content
@@ -27,11 +28,11 @@ class Webmirror extends AbstractSpider
      */
     protected array $body_save_listeners = [];
 
-    public function __construct(string $output_dir, HttpClientQueue $clientQueue, string $link_prefix = '')
+    public function __construct(string $output_dir, HttpClientQueue $clientQueue, string $path_prefix = '')
     {
         parent::__construct($clientQueue);
         $this->output_dir = $output_dir;
-        $this->link_prefix = $link_prefix;
+        $this->path_prefix = $path_prefix;
         $this->addUrlRewriter([$this, 'rewrite_links']);
     }
 
@@ -49,7 +50,6 @@ class Webmirror extends AbstractSpider
         if (!file_exists($this->output_dir)) {
             mkdir($this->output_dir, 0777, true);
         }
-        copy(__DIR__ . '/not_mirrored.php', $this->output_dir. '/not_mirrored.php');
 
         $this->clientQueue->start();
     }
@@ -72,41 +72,62 @@ class Webmirror extends AbstractSpider
         $this->save($request_url, $response);
     }
 
-    public function rewrite_links($accepted, $url)
+    public function rewrite_links($accepted, string $original_url, UriInterface $rewritten_url = null)
     {
-        if (!$url instanceof UriInterface) {
-            $url = Utils::uriFor($url);
-            if ($url->getScheme() == 'data' || $url->getScheme() == 'mailto' || $url->getScheme() == 'tel') {
-                return $url;
+        $url = Utils::uriFor($original_url);
+        if ($url->getScheme() == 'data' || $url->getScheme() == 'mailto' || $url->getScheme() == 'tel') {
+            return $url;
+        }
+        if ($accepted && $rewritten_url) {
+            if ($rewritten_url->getHost() == $this->hostname) {
+                $rewritten_url = $rewritten_url->withScheme('')->withHost('');
             }
-        }
-        if (!$accepted) {
-            return $this->link_prefix . '/not_mirrored.php?url=' . urlencode((string) $url);
-        }
-        if ($url instanceof UriInterface) {
-            $url = $url->getPath();
-        }
-        if (substr($url, 0, 1) === '/') {
-            $url = $this->link_prefix . $url;
+
+            # merge query hash into path
+            if ($rewritten_url->getQuery()) {
+                $rewritten_url = $rewritten_url->withPath($this->compute_path_for_uri($rewritten_url));
+                $rewritten_url = $rewritten_url->withQuery('');
+            }
+
+            # re-add fragment from original url
+            if ($url->getFragment()) {
+                $rewritten_url = $rewritten_url->withFragment($url->getFragment());
+            }
+
+            //            if (substr($url, 0, 1) === '/') {
+//                $url = $this->link_prefix . $url;
+//            }
+            if ($this->path_prefix) {
+                $rewritten_url = $rewritten_url->withPath($this->path_prefix . $rewritten_url->getPath());
+            }
+
+            $url = $rewritten_url;
         }
         return $url;
     }
 
-    protected function correct_path($path)
+    /**
+     * Compute a file path for saving.
+     * If there is no filename with an extension, add /index.html.
+     * Merge query hash into file path.
+     *
+     * @param UriInterface $uri
+     * @return string
+     */
+    protected function compute_path_for_uri(UriInterface $uri)
     {
+        $query = $uri->getQuery();
+        $path = $uri->getPath();
         $filename = basename($path);
         if (substr($path, -1) == '/') {
             $path .= 'index.html';
         } else if (strpos($filename, '.') === false) {
             $path = $path . '/index.html';
         }
-        return $path;
-    }
-
-    protected function prepare_fs_path($request_path)
-    {
-        $path = $this->correct_path($request_path);
-        $path = $this->output_dir . $path;
+        if ($query) {
+            $extension = pathinfo($path, PATHINFO_EXTENSION);
+            $path = $path . '.' . md5($query) . '.' . $extension;
+        }
         return $path;
     }
 
@@ -116,13 +137,21 @@ class Webmirror extends AbstractSpider
      */
     protected function save($url, $response)
     {
-        $path = $this->prepare_fs_path(Utils::uriFor($url)->getPath());
+        $path = $this->output_dir . $this->compute_path_for_uri(Utils::uriFor($url));
         $dir = dirname($path);
         if (!file_exists($dir)) {
             mkdir($dir, 0777, true);
         }
         $save = true;
         $last_modified = $response->getHeader('last-modified');
+        $md5_response = md5((string) $response->getBody());
+        if (empty($this->hashes[$md5_response])) {
+            $this->hashes[$md5_response] = [];
+        }
+        if (!in_array($path, $this->hashes[$md5_response])) {
+            $this->hashes[$md5_response][] = $path;
+        }
+
         if (!empty($last_modified)) {
             $last_modified = (is_array($last_modified) ? $last_modified[0] : $last_modified);
             $last_modified = new \DateTimeImmutable($last_modified);
@@ -134,14 +163,26 @@ class Webmirror extends AbstractSpider
                 if ($last_modified <= $filemtime) {
                     $save = false;
                 }
+            } else {
+                $md5_existing = md5_file($path);
+                if ($md5_existing === $md5_response) {
+                    $save = false;
+                    $this->logger->debug(sprintf('%s not saved because checksum has not changed', $path));
+                }
             }
         }
         if ($save) {
-            file_put_contents($path, ($response instanceof ResponseInterface ? $response->getBody() : $response));
+            $firstpath = $this->hashes[$md5_response][0];
+            if (count($this->hashes[$md5_response]) > 1 && file_exists($firstpath)) {
+                link($firstpath, $path);
+                $this->logger->info(sprintf('%s created as link to %s because of identical content', $path, $firstpath));
+            } else {
+                file_put_contents($path, ($response instanceof ResponseInterface ? $response->getBody() : $response));
+                if ($this->logger) $this->logger->info('File saved: ' . $path);
+            }
             if (!empty($last_modified)) {
                 touch($path, $last_modified);
             }
-            if ($this->logger) $this->logger->info('File saved: ' . $path);
         } else {
             if ($this->logger) $this->logger->info('File already exists: ' . $path);
         }
